+++ ST/AlyonaIpatova/Project/src/components/index.js
@@ -1,10 +1,6 @@
-import Basket from './basket';
-import Catalog from './catalog';
+import { Basket, Catalog } from './LISTS';
 
 export default () => {
-    let urlBasket = 'https://raw.githubusercontent.com/kellolo/static/master/JSON/basket.json';
-    let urlCatalog = 'https://raw.githubusercontent.com/kellolo/static/master/JSON/catalog.json';
-
-    let basket = new Basket ('#basket', urlBasket);
-    let catalog = new Catalog ('#catalog', basket, urlCatalog)
+    let basket = new Basket();
+    let catalog = new Catalog(basket);
 }
